+++ src/components/dropdownSearch/MultiSelectDropdownSearch/MultiSelectDropdownSearchConnected.tsx
@@ -3,6 +3,13 @@ import * as _ from 'underscore';
 import {IReactVaporState, IReduxActionsPayload} from '../../../ReactVapor';
 import {IReduxAction, ReduxUtils} from '../../../utils/ReduxUtils';
 import {
+    IDropdownOption,
+    IDropdownSearchDispatchProps,
+    IDropdownSearchOwnProps,
+    IDropdownSearchProps,
+    IDropdownSearchStateProps,
+} from '../DropdownSearch';
+import {
     addCustomSelectedOption,
     addMultiSelectDropdownSearch,
     applyFilterDropdownSearch,
@@ -15,14 +22,7 @@ import {
     removeDropdownSearch,
     updateOptionsDropdownSearch,
 } from '../DropdownSearchActions';
-import {
-    IDropdownOption,
-    IDropdownSearchDispatchProps,
-    IDropdownSearchOwnProps,
-    IDropdownSearchProps,
-    IDropdownSearchStateProps,
-} from './../DropdownSearch';
-import {IDropdownSearchState} from './../DropdownSearchReducers';
+import {IDropdownSearchState} from '../DropdownSearchReducers';
 import {MultiSelectDropdownSearch} from './MultiSelectDropdownSearch';
 
 const mapStateToProps = (state: IReactVaporState, ownProps: IDropdownSearchProps): IDropdownSearchStateProps => {
