+++ src/components/input/tests/InputTestCommons.spec.ts
@@ -1,4 +1,4 @@
-import {IInputProps} from './../Input';
+import {IInputProps} from '../Input';
 
 export const inputProps: IInputProps = {
     id: 'test-input',
